requirements.txt
aiohttp==3.6.1
PyJWT==1.7.1
sentry-sdk==0.12.2
envparse==0.2.0
python-magic==0.4.15
mutagen==1.42.0
Werkzeug==0.16.0
aioredis==1.3.0
aiofiles==0.4.0
Pillow==6.2.0
pypng==0.0.20
aiohttp-cors==0.7.0
aiobotocore==0.11.1